requirements.txt
flask
torch
torchvision
torchaudio
transformers
Pillow
